Rename ChatInputx and ChatListx files to ChatInput and ChatList respectively, and update imports in Chat/index.tsx file
